Recompute envelope curves
Original envelope curves, included in the data set published with the paper, do not
always seem to reach to the end of the load history. Some also seem erroneous in other ways. For this reason, all envelope curves were recomputed. The Matlab files for recomputing the envelope curves are included in the repository (subfolder 01_Matlab_scripts).